5 dec 2025

+++ data_procedure/step1/face_motion_and_trimming.py
@@ -22,8 +22,10 @@ abs_r = True
 max_lag_frames = 88 # About 5 seconds at 17.5 fps
 lag_range = (-max_lag_frames, max_lag_frames)
 
+skip_face_motion = False
 
-mice_num = ["39-12", "42-12", "44-12", "45-12", "46-12", "251-6", "254-6"]
+# "308-16", "316-14", "316-16", "322-14", 
+mice_num = ["308-10"]
 
 for mouse_num in mice_num:
 
@@ -32,48 +34,55 @@ for mouse_num in mice_num:
     # COMPUTE FACE MOTION #
     ### FACE MOTION CODE MADE BY ANTOINE DAIGLE
 
-    video_input, video_output, face_motion_output, HbT_input, datafile_input = return_filepaths(mouse_num)
-    print(f"Opening video {video_input}.")
+    if not skip_face_motion:
 
-    cap = cv2.VideoCapture(video_input)
-    video_length = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+        video_input, video_output, face_motion_output, cortical_signal_input, datafile_input = return_filepaths(mouse_num)
+        print(f"Opening video {video_input}.")
 
-    # Check if the video opened successfully
-    if not cap.isOpened():
-        print("Error: Could not open video.")
-        exit()
+        cap = cv2.VideoCapture(video_input)
+        video_length = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
 
-    ret, frame = cap.read()
-    gray_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+        # Check if the video opened successfully
+        if not cap.isOpened():
+            print("Error: Could not open video.")
+            exit()
 
-    x_mot, y_mot, w_mot, h_mot = cv2.selectROI("Select motion zone", gray_frame)
-
-    motion = np.zeros(shape=video_length)
-
-    for i in tqdm(range(video_length)):
-        # Read frame-by-frame
-        cap.set(cv2.CAP_PROP_POS_FRAMES, i)
         ret, frame = cap.read()
-
-        # Convert the frame to grayscale
         gray_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 
-        motion_zone = gray_frame[y_mot:y_mot+h_mot, x_mot:x_mot+w_mot]
+        x_mot, y_mot, w_mot, h_mot = cv2.selectROI("Select motion zone", gray_frame)
 
-        if i == 0:
-            old_motion_zone = motion_zone.copy()
+        motion = np.zeros(shape=video_length)
 
-        else:
-            motion[i] = cv2.mean(cv2.absdiff(motion_zone, old_motion_zone))[0]
-            old_motion_zone = motion_zone.copy()
+        for i in tqdm(range(video_length)):
+            # Read frame-by-frame
+            cap.set(cv2.CAP_PROP_POS_FRAMES, i)
+            ret, frame = cap.read()
 
-    # Release the video capture object and close the display window
-    cap.release()
-    cv2.destroyAllWindows()
+            # Convert the frame to grayscale
+            gray_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
 
-    np.save(face_motion_output, motion)
+            motion_zone = gray_frame[y_mot:y_mot+h_mot, x_mot:x_mot+w_mot]
 
-    print(f"Face motion saved.")
+            if i == 0:
+                old_motion_zone = motion_zone.copy()
+
+            else:
+                motion[i] = cv2.mean(cv2.absdiff(motion_zone, old_motion_zone))[0]
+                old_motion_zone = motion_zone.copy()
+
+        # Release the video capture object and close the display window
+        cap.release()
+        cv2.destroyAllWindows()
+
+        np.save(face_motion_output, motion)
+
+        print(f"Face motion saved.")
+
+    elif skip_face_motion:
+
+        video_input, video_output, face_motion_output, cortical_signal_input, datafile_input = return_filepaths(mouse_num)
+        face_motion_signal = np.load(face_motion_output)
 
     # TRIM VIDEO AND FACE MOTION TO APPROPRIATE FPS
 
@@ -94,20 +103,22 @@ for mouse_num in mice_num:
     # Load datasets
 
     # TODO: the only issue here is that it considers the whole HbT window, not just the ROI since it loads the raw file. Maybe to fix eventually but should be similar in any case.
-    HbT_signal = read_data(HbT_input)
-    duration = HbT_signal.shape[0] # At 12 fps, should be 5760 (8 minutes). 1440 for 3 fps.
+    cortical_signal = read_data(cortical_signal_input)
+    duration = cortical_signal.shape[0] # At 12 fps, should be 5760 (8 minutes). 1440 for 3 fps.
     print(f"HbT data with {duration} frames loaded. Trimming the video and face motion to match.")
 
-    HbT_signal = np.nanmean(HbT_signal, axis=(1, 2))
+    cortical_signal = np.nanmean(cortical_signal, axis=(1, 2))
 
     face_motion_signal = np.load(face_motion_output)
 
-    # Find the indices that correctly spread the monitoring frames across the HbT timeseries
+    # Find the indices that correctly spread the monitoring frames across the cortical (HbT/GCaMP) timeseries
     indices_vec = np.arange(frame_range[0], frame_range[1], 1)
+    # print(indices_vec)
     indices = evenly_spread_vector_elements(indices_vec, duration)
+    # print(indices)
 
-    # Cross correlate face_motion with HbT
-    correlations = compute_correlation_with_lag(HbT_signal, face_motion_signal, indices, lag_range, abs_r=abs_r)
+    # Cross correlate face_motion with the cortical signal (HbT or GCaMP)
+    correlations = compute_correlation_with_lag(cortical_signal, face_motion_signal, indices, lag_range, abs_r=abs_r)
     max_correl_index = np.argmax(correlations)
     max_correl_lag = np.arange(lag_range[0], lag_range[1]+1, 1)[max_correl_index]
